+++ lib/application.dart
@@ -1,0 +1,4 @@
+library;
+
+export 'src/application/state.dart';
+export 'src/application/state_notifier.dart';