+++ lib/b_router.dart
@@ -1,6 +1,6 @@
 library b_router;
 
 export 'core.dart';
-export 'bloc.dart';
 export 'router.dart';
 export 'widgets.dart';
+export 'application.dart';