+++ lib/bloc.dart
@@ -1,4 +1,0 @@
-/// Contains the bloc related files.
-library bloc;
-
-export 'src/bloc/b_router_cubit.dart';