+++ lib/router.dart
@@ -1,6 +1,8 @@
 /// Library which contains router related classes.
 library router;
 
-export 'src/router/route.dart';
 export 'src/router/delegate.dart';
+export 'src/router/information_provider.dart';
 export 'src/router/parser.dart';
+export 'src/router/route_event.dart';
+export 'src/router/route.dart';