+++ lib/src/core/extensions.dart
@@ -1,27 +1,41 @@
+import 'package:b_router/src/router/route_event.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:b_router/bloc.dart';
+import '../application/state.dart';
 import '../router/route.dart';
+import '../widgets/state_provider.dart';
 
 extension BRouterContextExtensions on BuildContext {
-  /// Get the nearest [BRouterCubit] instance.
-  ///
-  /// Use [watchBRouter] if you need to retrieve it inside a build method of a widget.
-  BRouterCubit get bRouter => read<BRouterCubit>();
+  /// Get the current BRouteStateProvider from the context.
+  BRouterStateProvider get routerStateProvider => BRouterStateProvider.of(this);
 
-  /// Watch for [BRouterCubit] changes inside build methods.
+  /// Get the current [BRouterState] from the context.
+  BRouterState get routerState => routerStateProvider.state;
+
+  /// Get the currently pushed routes from the context.
   ///
-  /// For one time usage use [bRouter].
-  BRouterCubit get watchBRouter => watch<BRouterCubit>();
+  /// Keep in mind that this will return an empty list if the current state does not contain any
+  /// routes.
+  List<BRoute> get pushedRoutes => switch (routerState) {
+        FoundRoutes(:final routes) => routes,
+        _ => const [],
+      };
+
+  /// Get the last route pushed.
+  ///
+  /// This will return null if there are no routes in the stack.
+  ///
+  /// It's a shorthand for `pushedRoutes.lastOrNull`.
+  BRoute? get topRoute => pushedRoutes.lastOrNull;
 
   /// Redirect to the new [location].
   ///
   /// For more see [BRouterCubit.redirect].
-  void bRedirect({required String location, Map<String, String>? params}) => bRouter.redirect(
-        location: location,
-        params: params,
-      );
+  void bRedirect({
+    required String location,
+    Map<String, String>? params,
+  }) =>
+      RedirectRouteEvent(location: location, params: params).dispatch(this);
 
   /// Push a new page to the stack.
   ///
@@ -31,15 +45,5 @@ extension BRouterContextExtensions on BuildContext {
     Map<String, dynamic>? arguments,
     Map<String, String>? params,
   }) =>
-      bRouter.push(
-        name: name,
-        arguments: arguments,
-        params: params,
-      );
-
-  /// Get the last route pushed.
-  BRoute? get bTopRoute => switch (bRouter.state) {
-        FoundRoutes(:final routes) => routes.isNotEmpty ? routes.last : null,
-        _ => null,
-      };
+      PushRouteEvent(name: name, arguments: arguments, params: params).dispatch(this);
 }